pyproject.toml
[project]
name = "slidex"
version = "0.1.0"
description = "Slidex - Local PowerPoint slide management with semantic search"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "python-pptx>=1.0.2",
    "pillow>=11.0.0",
    "faiss-cpu>=1.9.0",
    "flask>=3.1.0",
    "jinja2>=3.1.4",
    "typer>=0.15.1",
    "pydantic>=2.10.5",
    "pydantic-settings>=2.7.0",
    "psycopg2-binary>=2.9.10",
    "loguru>=0.7.3",
    "ollama>=0.4.5",
    "pyyaml>=6.0.2",
    "numpy>=2.2.1",
    "pytest>=8.3.4",
]

[project.scripts]
slidex = "slidex.cli.main:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
